+++ src/esm/drawutilssvg.js
@@ -50,6 +50,7 @@
  * @modified 2025-11-14 Fixing a bug in the CSS `mix-blend-mode` property handling (caused a runtime error).
  * @modified 2026-01-04 Adding `lineJoin` attribute to the methods' `StrokeOptions` param.
  * @modified 2026-01-04 Fixing missing `strokeOptions` param in the `drawutilssvg.polygon` method.
+ * @modified 2026-03-18 Adding `isOpen` parameter to `cubicBezierPath` draw method.
  * @version  1.7.0
  **/
 import { CircleSector } from "./CircleSector";
@@ -723,12 +724,12 @@ export class drawutilssvg {
      * @param {string} color - The CSS colot to draw the path with.
      * @param {number=1} lineWidth - (optional) The line width to use.
      * @param {StrokeOptions=} strokeOptions - (optional) Stroke settings to use.
-     *
+     * @param {boolean} isOpen - (optional) Set to tur if the path should not be closed.
      * @return {void}
      * @instance
      * @memberof drawutilssvg
      */
-    cubicBezierPath(path, color, lineWidth, strokeOptions) {
+    cubicBezierPath(path, color, lineWidth, strokeOptions, isOpen) {
         const node = this.makeNode("path");
         this.applyStrokeOpts(node, strokeOptions);
         if (!path || path.length == 0) {
@@ -745,6 +746,9 @@ export class drawutilssvg {
             endControlPoint = path[i + 1];
             endPoint = path[i + 2];
             d.push("C", this._x(startControlPoint.x), this._y(startControlPoint.y), this._x(endControlPoint.x), this._y(endControlPoint.y), this._x(endPoint.x), this._y(endPoint.y));
+        }
+        if (!isOpen) {
+            d.push("Z");
         }
         node.setAttribute("d", d.join(" "));
         return this._bindFillDraw(node, "cubicBezierPath", color, lineWidth || 1);