+++ src/ts/draw.ts
@@ -58,6 +58,7 @@
  * @modified 2023-10-07 Adding the optional `arrowHeadBasePositionBuffer` param to the arrowHead(...) method.
  * @modified 2024-09-13 Remoed the scaling of `lineWidth` in the `polygon` and `polyline` methods. This makes no sense here and doesn't match up with the behavior of other line functions.
  * @modified 2026-01-04 Adding `lineJoin` attribute to the `StrokeOptions`.
+ * @modified 2026-03-18 Adding `isOpen` parameter to `cubicBezierPath` draw method.
  * @version  1.14.0
  **/
 
@@ -700,7 +701,7 @@ export class drawutils implements DrawLib<void> {
    * @instance
    * @memberof drawutils
    */
-  cubicBezierPath(path: Array<XYCoords>, color: string, lineWidth?: number, strokeOptions?: StrokeOptions) {
+  cubicBezierPath(path: Array<XYCoords>, color: string, lineWidth?: number, strokeOptions?: StrokeOptions, isOpen?: boolean) {
     if (!path || path.length == 0) {
       return;
     }
@@ -725,7 +726,9 @@ export class drawutils implements DrawLib<void> {
         this.offset.y + endPoint.y * this.scale.y
       );
     }
-    this.ctx.closePath();
+    if (!isOpen) {
+      this.ctx.closePath();
+    }
     this.ctx.lineWidth = lineWidth || 1;
     this._fillOrDraw(color);
     this.ctx.restore();